Fix: Keep state persistent when changing pages on return request form
